Added test script to see which functions still need to be implemented.

+++ src/dcopy.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-float* dcopy(size_t N, double* X, int INCX, double* Y, int INCY){
+double* dcopy(size_t N, double* X, int INCX, double* Y, int INCY){
     int d = 0, ix = 0, iy = 0;
     for(d = 0; d < N; d++){
         Y[iy] = X[ix];
