+++ src/dscal.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-float* dscal(size_t N, double ALPHA, double* X, int INCX){
+double* dscal(size_t N, double ALPHA, double* X, int INCX){
     int d = 0, ix = 0;
     for(d = 0; d < N; d++){
         X[ix] *= ALPHA;
